Particle system and Geometry overhaul
Re-implemented bounding box intersection with normals
Added convenience Point::subtract
Renamed scale() to multiply()
Introduced Intersects abstraction
Refactored intersection for lazy normals
Added min/max to Tuple
Added divisor() for vector division
Added creation time to particles, implemented culling
Refactored particle system for parallel iteration, added update helper
Updated particle system doc
Added vector randomiser utility

+++ src/main/java/org/sarge/jove/particle/Influence.java
@@ -10,9 +10,9 @@ public interface Influence {
 	/**
 	 * Applies this influence to the given particle.
 	 * @param p 			Particle to influence
-	 * @param elapsed		Elapsed time since previous update (ms)
+	 * @param elapsed		Elapsed scalar
 	 */
-	void apply(Particle p, long elapsed);
+	void apply(Particle p, float elapsed);
 
 	/**
 	 * Creates an acceleration influence by incrementing the particle direction by the given vector.
